Create PreTypeCheck pass
This moves more checks that don't involve types before the type resolving
and checking passes. It adds more visitors to the variable_precheck and
changes the name of this pass to be more general: pre_type_check.

The primary motivation is to have error messages surface earlier so we
don't have to deal with certain error path scenarios (like function
argument count validation) in type passes.

stack-info: PR: https://github.com/bpftrace/bpftrace/pull/5001, branch: user/jordalgo/pre_type_check/1
