+++ tests/CMakeLists.txt
@@ -11,6 +11,7 @@ add_executable(bpftrace_test
   bpftrace.cpp
   btf.cpp
   builtins.cpp
+  pre_type_check.cpp
   child.cpp
   clang_parser.cpp
   named_param.cpp
@@ -51,7 +52,6 @@ add_executable(bpftrace_test
   type_system.cpp
   unstable_feature.cpp
   utils.cpp
-  variable_precheck.cpp
 )
 add_test(NAME bpftrace_test COMMAND bpftrace_test)
 file(WRITE ${CMAKE_CURRENT_BINARY_DIR}/near_self_file "near_self_file")
